+++ src/packerlicious/post_processor.py
@@ -1,4 +1,5 @@
 from . import BasePackerObject
+import validator
 
 
 class PackerPostProcessor(BasePackerObject):
@@ -17,4 +18,24 @@ class DockerImport(PackerPostProcessor):
     props = {
         'repository': (basestring, True),
         'tag': (basestring, False),
+    }
+
+
+class DockerPush(PackerPostProcessor):
+    """
+    Docker Push Processor
+    https://www.packer.io/docs/post-processors/docker-push.html
+    """
+    resource_type = "docker-push"
+
+    props = {
+        'aws_access_key': (basestring, False),
+        'aws_secret_key': (basestring, False),
+        'aws_token': (basestring, False),
+        'ecr_login': (validator.boolean, False),
+        'login': (validator.boolean, False),
+        'login_email': (basestring, False),
+        'login_username': (basestring, False),
+        'login_password': (basestring, False),
+        'login_server': (basestring, False),
     }