+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 * builder/docker: add packer docker builder support
 * builder/file: add packer file builder support
 * provisioner/DockerImport: add packer docker import provisioner support
+* provisioner/DockerPush: add packer docker push provisioner support
 * provisioner/file: add packer file provisioner support